+++ syncdata/operation/list_lister_test.go
@@ -1,4 +1,4 @@
-package test
+package operation
 
 import (
 	"context"
