+++ syncdata/operation/utils.go
@@ -1,5 +1,7 @@
 package operation
 
+import "strings"
+
 func dupStrings(s []string) []string {
 	if len(s) == 0 {
 		return s
@@ -7,4 +9,11 @@ func dupStrings(s []string) []string {
 	to := make([]string, len(s))
 	copy(to, s)
 	return to
+}
+
+func makeSureKeyAsDir(key string) string {
+	if strings.HasSuffix(key, "/") {
+		return key
+	}
+	return key + "/"
 }
